# docs/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cultivar Collection Management API",
    "description": "Advanced cannabis cultivar tracking and analytics platform",
    "version": "2.0.0",
    "x-generated": {
      "timestamp": "2025-11-01T19:55:58.334401",
      "source": "app.fastapi_app",
      "generator": "scripts/generate_openapi.py"
    }
  },
  "paths": {
...
    "/plants/": {
      "get": {
        "tags": [
          "Plants Legacy",
          "plants LEGACY - DEPRECATED - Use /api/v1/plants"
        ],
        "summary": "Plants List",
        "description": "DEPRECATED: Use /api/v1/plants/list instead - Display list of user's plants",
        "operationId": "plants_list_plants__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "deprecated": true,
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/plants/{plant_id}": {
      "get": {
        "tags": [
          "Plants Legacy",
          "plants LEGACY - DEPRECATED - Use /api/v1/plants"
        ],
        "summary": "Plant Detail",
        "description": "DEPRECATED: Use /api/v1/plants/{plant_id} instead - Display plant detail page.",
        "operationId": "plant_detail_plants__plant_id__get",
        "deprecated": true,
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/plants/new": {
      "get": {
        "tags": [
          "Plants Legacy",
          "plants LEGACY - DEPRECATED - Use /api/v1/plants"
        ],
        "summary": "Plant Create Page",
        "description": "DEPRECATED: Use /api/v1/plants (POST) instead - Display plant creation form.",
        "operationId": "plant_create_page_plants_new_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "deprecated": true,
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/plants/{plant_id}/edit": {
      "get": {
        "tags": [
          "Plants Legacy",
          "plants LEGACY - DEPRECATED - Use /api/v1/plants"
        ],
        "summary": "Plant Edit Page",
        "description": "DEPRECATED: Use /api/v1/plants/{plant_id} (PATCH) instead - Display plant edit form.",
        "operationId": "plant_edit_page_plants__plant_id__edit_get",
        "deprecated": true,
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/plants/api/v1/plants/": {
      "get": {
        "tags": [
          "Plants API",
          "plants"
        ],
        "summary": "List Plants",
        "description": "Get paginated list of user's plants",
        "operationId": "list_plants_api_v1_plants_api_v1_plants__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 20,
              "title": "Page Size"
            },
            "description": "Items per page"
          },
          {
            "name": "status_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status ID",
              "title": "Status Id"
            },
            "description": "Filter by status ID"
          },
          {
            "name": "cultivar_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by cultivar ID",
              "title": "Cultivar Id"
            },
            "description": "Filter by cultivar ID"
          },
          {
            "name": "is_clone",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by clone status",
              "title": "Is Clone"
            },
            "description": "Filter by clone status"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search in plant names",
              "title": "Search"
            },
            "description": "Search in plant names"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlantListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Plants API",
          "plants"
        ],
        "summary": "Create Plant",
        "description": "Create a new plant",
        "operationId": "create_plant_api_v1_plants_api_v1_plants__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlantCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlantResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/plants/api/v1/plants/{plant_id}": {
      "get": {
        "tags": [
          "Plants API",
          "plants"
        ],
        "summary": "Get Plant",
        "description": "Get a single plant by ID",
        "operationId": "get_plant_api_v1_plants_api_v1_plants__plant_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlantResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Plants API",
          "plants"
        ],
        "summary": "Update Plant",
        "description": "Update an existing plant",
        "operationId": "update_plant_api_v1_plants_api_v1_plants__plant_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plant Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__fastapi_app__routers__plants_api__PlantUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlantResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Plants API",
          "plants"
        ],
        "summary": "Delete Plant",
        "description": "Delete a plant",
        "operationId": "delete_plant_api_v1_plants_api_v1_plants__plant_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plant Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/stats": {
      "get": {
        "tags": [
          "Dashboard",
          "dashboard"
        ],
        "summary": "Dashboard Stats",
        "description": "Get comprehensive dashboard statistics for the current user.\n\nThis endpoint provides all the data needed to populate the Flutter dashboard\nwith real-time information including counts, environmental data, and charts.",
        "operationId": "dashboard_stats_stats_get",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DashboardStatsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardStats"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/plants": {
      "get": {
        "tags": [
          "Dashboard",
          "dashboard"
        ],
        "summary": "Dashboard Plants",
        "description": "Plants page.",
        "operationId": "dashboard_plants_plants_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DashboardStatsRequest": {
        "properties": {
          "include_history": {
            "type": "boolean",
            "title": "Include History",
            "description": "Include historical data for charts",
            "default": true
          },
          "time_range_hours": {
            "type": "integer",
            "title": "Time Range Hours",
            "description": "Time range for historical data in hours",
            "default": 24
          },
          "limit_recent_readings": {
            "type": "integer",
            "title": "Limit Recent Readings",
            "description": "Limit number of recent readings",
            "default": 10
          },
          "limit_recent_activities": {
            "type": "integer",
            "title": "Limit Recent Activities",
            "description": "Limit number of recent activities",
            "default": 10
          }
        },
        "type": "object",
        "title": "DashboardStatsRequest",
        "description": "Request model for dashboard stats with optional parameters."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlantCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "description": "Plant name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Plant description"
          },
          "status_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Status Id",
            "description": "Status ID"
          },
          "cultivar_id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Cultivar Id",
            "description": "Cultivar ID"
          },
          "is_clone": {
            "type": "boolean",
            "title": "Is Clone",
            "description": "Whether this is a clone",
            "default": false
          },
          "autoflower": {
            "type": "boolean",
            "title": "Autoflower",
            "description": "Whether this is an autoflower variety",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name",
          "status_id"
        ],
        "title": "PlantCreate",
        "description": "Request model for creating a plant"
      },
      "PlantListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PlantResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next"
          },
          "has_prev": {
            "type": "boolean",
            "title": "Has Prev"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "has_next",
          "has_prev"
        ],
        "title": "PlantListResponse",
        "description": "Response model for plant list with pagination"
      },
      "PlantResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status_id": {
            "type": "integer",
            "title": "Status Id"
          },
          "cultivar_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cultivar Id"
          },
          "is_clone": {
            "type": "boolean",
            "title": "Is Clone"
          },
          "autoflower": {
            "type": "boolean",
            "title": "Autoflower"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "last_water_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Water Date"
          },
          "last_feed_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Feed Date"
          },
          "harvest_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Harvest Date"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "status_id",
          "cultivar_id",
          "is_clone",
          "autoflower",
          "start_date",
          "last_water_date",
          "last_feed_date",
          "harvest_date",
          "user_id",
          "created_at",
          "updated_at"
        ],
        "title": "PlantResponse",
        "description": "Response model for a single plant"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}